+++ samplesheets/plugins.py
@@ -177,6 +177,26 @@ class ProjectAppPlugin(
             'access, used with projects allowing public guest access',
             'user_modifiable': False,
         },
+        'igv_omit_bam': {
+            'scope': SODAR_CONSTANTS['APP_SETTING_SCOPE_PROJECT'],
+            'type': 'STRING',
+            'label': 'BAM files to omit from IGV sessions',
+            'default': '',
+            'description': 'Comma separated list of BAM file suffixes to omit '
+            'from generated IGV sessions. Overrides site-wide '
+            'setting.',
+            'user_modifiable': True,
+        },
+        'igv_omit_vcf': {
+            'scope': SODAR_CONSTANTS['APP_SETTING_SCOPE_PROJECT'],
+            'type': 'STRING',
+            'label': 'VCF files to omit from IGV sessions',
+            'default': '',
+            'description': 'Comma separated list of VCF file suffixes to omit '
+            'from generated IGV sessions. Overrides site-wide '
+            'setting.',
+            'user_modifiable': True,
+        },
     }
 
     #: Iconify icon